Doradjena GrupeForma, dodati DTOs za Grupe i funkcije za vracanje grupa iz baze u DTOManager fajlu

+++ Aplikacija/SBP-faza2/SBP-faza2/Entiteti/Grupa.cs
@@ -2,7 +2,7 @@
 
 public class Grupa
 {
-    public virtual int Id { get; protected set; }
+    public virtual long Id { get; protected set; }
     public virtual required Projekat Projekat { get; set; }
     public virtual required string NazivGrupe { get; set; }
     public virtual IList<Izvestaj> Izvestaji { get; set; }
